fix(cursor): split daemon helpers to satisfy maintainability gate
